Move run_args and run_paths fixtures to unit_tests
